chore: jsx확장자에서 eslint오류가 발생하여 tsx로 변경
